feat: Add new hooks for escrow dialog functionalities including change flag, change status, edit properties, edit entities, edit milestones, fund escrow, release funds, and dispute resolution
