+++ src/components/modules/escrow/hooks/change-flag-escrow-dialog.hook.ts
@@ -6,9 +6,9 @@ import {
   useGlobalAuthenticationStore,
   useGlobalBoundedStore,
 } from "@/core/store/data";
-import { useEscrowUIBoundedStore } from "../../../store/ui";
+import { useEscrowUIBoundedStore } from "../store/ui";
 import { Escrow, Milestone } from "@/@types/escrows/escrow.entity";
-import { trustlessWorkService } from "../../../services/trustless-work.service";
+import { trustlessWorkService } from "../services/trustless-work.service";
 import { EscrowRequestResponse } from "@/@types/escrows/escrow-response.entity";
 import { ChangeMilestoneFlagPayload } from "@/@types/escrows/escrow-payload.entity";
 import { toast } from "sonner";
