+++ src/components/modules/escrow/hooks/edit-entities-dialog.hook.ts
@@ -9,28 +9,40 @@ import {
   useGlobalAuthenticationStore,
   useGlobalBoundedStore,
 } from "@/core/store/data";
-import { formSchema } from "../../../schema/edit-milestone.schema";
-import { useEscrowUIBoundedStore } from "../../../store/ui";
-import { Milestone } from "@/@types/escrows/escrow.entity";
+import { useEscrowUIBoundedStore } from "../store/ui";
+import { useMemo, useState } from "react";
+import { GetFormSchema } from "../schema/edit-entities.schema";
 import {
   EscrowPayload,
   UpdateEscrowPayload,
 } from "@/@types/escrows/escrow-payload.entity";
-import { trustlessWorkService } from "../../../services/trustless-work.service";
+import { trustlessWorkService } from "../services/trustless-work.service";
 import { EscrowRequestResponse } from "@/@types/escrows/escrow-response.entity";
 import { toast } from "sonner";
+import { useUsersQuery } from "@/components/modules/contact/hooks/tanstack/useUsersQuery";
 
-interface useEditMilestonesDialogProps {
-  setIsEditMilestoneDialogOpen: (value: boolean) => void;
+interface useEditEntitiesDialogProps {
+  setIsEditEntitiesDialogOpen: (value: boolean) => void;
 }
 
-const useEditMilestonesDialog = ({
-  setIsEditMilestoneDialogOpen,
-}: useEditMilestonesDialogProps) => {
+const useEditEntitiesDialog = ({
+  setIsEditEntitiesDialogOpen,
+}: useEditEntitiesDialogProps) => {
+  const formSchema = GetFormSchema();
+
+  const [showSelect, setShowSelect] = useState({
+    approver: false,
+    serviceProvider: false,
+    platformAddress: false,
+    releaseSigner: false,
+    disputeResolver: false,
+    receiver: false,
+  });
+
   const { address } = useGlobalAuthenticationStore();
   const selectedEscrow = useGlobalBoundedStore((state) => state.selectedEscrow);
-  const setIsEditingMilestones = useEscrowUIBoundedStore(
-    (state) => state.setIsEditingMilestones,
+  const setIsEditingEntities = useEscrowUIBoundedStore(
+    (state) => state.setIsEditingEntities,
   );
   const fetchAllEscrows = useGlobalBoundedStore(
     (state) => state.fetchAllEscrows,
@@ -39,44 +51,48 @@ const useEditMilestonesDialog = ({
   const setIsDialogOpen = useEscrowUIBoundedStore(
     (state) => state.setIsDialogOpen,
   );
+  const { data: users = [] } = useUsersQuery();
+
+  const userOptions = useMemo(() => {
+    const options = users.map((user) => ({
+      value: user.address,
+      label: `${user.firstName} ${user.lastName}`,
+    }));
+
+    return [{ value: "", label: "Select an User" }, ...options];
+  }, [users]);
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
-      milestones: selectedEscrow?.milestones || [{ description: "" }],
+      approver: selectedEscrow?.roles?.approver || "",
+      serviceProvider: selectedEscrow?.roles?.serviceProvider || "",
+      platformAddress: selectedEscrow?.roles?.platformAddress || "",
+      receiver: selectedEscrow?.roles?.receiver || "",
+      releaseSigner: selectedEscrow?.roles?.releaseSigner || "",
+      disputeResolver: selectedEscrow?.roles?.disputeResolver || "",
     },
     mode: "onChange",
   });
 
-  const milestones: Milestone[] = form.watch("milestones");
-  const isAnyMilestoneEmpty = milestones.some(
-    (milestone) => milestone.description === "",
-  );
-
-  const handleAddMilestone = () => {
-    const currentMilestones = form.getValues("milestones");
-    const updatedMilestones = [
-      ...currentMilestones,
-      { description: "", status: "pending" },
-    ];
-    form.setValue("milestones", updatedMilestones);
-  };
-
-  const handleRemoveMilestone = (index: number) => {
-    const currentMilestones = form.getValues("milestones");
-    const updatedMilestones = currentMilestones.filter((_, i) => i !== index);
-    form.setValue("milestones", updatedMilestones);
+  const toggleField = (field: string, value: boolean) => {
+    setShowSelect((prev) => ({ ...prev, [field]: value }));
   };
 
   const onSubmit = async (payload: z.infer<typeof formSchema>) => {
     if (!selectedEscrow) return;
 
-    setIsEditingMilestones(true);
+    setIsEditingEntities(true);
 
     try {
       const updatedEscrow = {
         ...JSON.parse(JSON.stringify(selectedEscrow)),
-        milestones: payload.milestones,
+        approver: payload.approver,
+        serviceProvider: payload.serviceProvider,
+        platformAddress: payload.platformAddress,
+        receiver: payload.receiver,
+        releaseSigner: payload.releaseSigner,
+        disputeResolver: payload.disputeResolver,
       };
 
       // Plain the trustline
@@ -106,11 +122,11 @@ const useEditMilestonesDialog = ({
 
       if (response.status === "SUCCESS") {
         fetchAllEscrows({ address, type: activeTab || "approver" });
-        setIsEditMilestoneDialogOpen(false);
+        setIsEditEntitiesDialogOpen(false);
         setIsDialogOpen(false);
 
         toast.success(
-          `You have edited the milestones of ${selectedEscrow.title}.`,
+          `You have edited the entities of ${selectedEscrow.title}.`,
         );
       }
     } catch (err) {
@@ -118,23 +134,22 @@ const useEditMilestonesDialog = ({
         err instanceof Error ? err.message : "An unknown error occurred",
       );
     } finally {
-      setIsEditingMilestones(false);
+      setIsEditingEntities(false);
     }
   };
 
   const handleClose = () => {
-    setIsEditMilestoneDialogOpen(false);
+    setIsEditEntitiesDialogOpen(false);
   };
 
   return {
-    onSubmit,
     form,
+    userOptions,
+    showSelect,
+    toggleField,
+    onSubmit,
     handleClose,
-    milestones,
-    handleAddMilestone,
-    handleRemoveMilestone,
-    isAnyMilestoneEmpty,
   };
 };
 
-export default useEditMilestonesDialog;
+export default useEditEntitiesDialog;
