+++ src/components/modules/escrow/ui/dialogs/CompleteMilestoneDialog.tsx
@@ -18,7 +18,7 @@ import {
 import TooltipInfo from "@/components/utils/ui/Tooltip";
 import { useEscrowUIBoundedStore } from "../../store/ui";
 import { PackageCheck } from "lucide-react";
-import useCompleteMilestoneDialogHook from "./hooks/change-status-escrow-dialog.hook";
+import useCompleteMilestoneDialogHook from "../../hooks/change-status-escrow-dialog.hook";
 import { useEscrowBoundedStore } from "../../store/data";
 import SkeletonCompleteMilestone from "./utils/SkeletonCompleteMilestone";
 import { Textarea } from "@/components/ui/textarea";
